+++ Scripts/FollowGameObject.hpp
@@ -7,7 +7,7 @@
 
 #include <Components/BehaviourScript.hpp>
 
-class FollowGameObject : public BehaviourScript{
+class FollowGameObject : public BehaviourScript {
 public:
 public:
     FollowGameObject(std::string tag) : BehaviourScript(), followGameObjectTag(tag) {}
@@ -16,7 +16,7 @@ public:
 
     virtual void onStart() override;
 
-    virtual void onUpdate(float deltaTime) override;
+    virtual void onUpdate(int deltaTime) override;
 
     std::unique_ptr<IComponent> clone() const override {
         return std::make_unique<FollowGameObject>(*this);
@@ -27,11 +27,11 @@ public:
         onStart();
     }
 
-    FollowGameObject(const FollowGameObject& other) :
-        BehaviourScript(other),
-        moveCameraMargin(other.moveCameraMargin),
-        followGameObjectTag(other.followGameObjectTag)
-        {}
+    FollowGameObject(const FollowGameObject &other) :
+            BehaviourScript(other),
+            moveCameraMargin(other.moveCameraMargin),
+            followGameObjectTag(other.followGameObjectTag) {}
+
 private:
     float moveCameraMargin = 30;
     std::string followGameObjectTag;
