+++ Scripts/UserInputMovement.cpp
@@ -8,19 +8,21 @@
 
 void UserInputMovement::onStart() {}
 
-void UserInputMovement::onUpdate(float deltaTime) {
-    auto& playerVelocityComponent = tryGetComponent<VelocityComponent>();
+void UserInputMovement::onUpdate(int deltaTime) {
+    auto &playerVelocityComponent = tryGetComponent<VelocityComponent>();
 
     if (InputManager::GetInstance().IsKeyPressed(KeyMap::w) || InputManager::GetInstance().IsKeyPressed(KeyMap::UP)) {
         playerVelocityComponent.velocity.setY(-velocity);
-    } else if (InputManager::GetInstance().IsKeyPressed(KeyMap::s) || InputManager::GetInstance().IsKeyPressed(KeyMap::DOWN)) {
+    } else if (InputManager::GetInstance().IsKeyPressed(KeyMap::s) ||
+               InputManager::GetInstance().IsKeyPressed(KeyMap::DOWN)) {
         playerVelocityComponent.velocity.setY(velocity);
     } else {
         playerVelocityComponent.velocity.setY(0);
     }
     if (InputManager::GetInstance().IsKeyPressed(KeyMap::a) || InputManager::GetInstance().IsKeyPressed(KeyMap::LEFT)) {
         playerVelocityComponent.velocity.setX(-velocity);
-    } else if (InputManager::GetInstance().IsKeyPressed(KeyMap::d) || InputManager::GetInstance().IsKeyPressed(KeyMap::RIGHT)) {
+    } else if (InputManager::GetInstance().IsKeyPressed(KeyMap::d) ||
+               InputManager::GetInstance().IsKeyPressed(KeyMap::RIGHT)) {
         playerVelocityComponent.velocity.setX(velocity);
     } else {
         playerVelocityComponent.velocity.setX(0);
